requirements.txt
beautifulsoup4==4.14.3
openai==2.30.0
pandas==3.0.2
python-dotenv==1.2.2
Requests==2.33.1
streamlit==1.56.0
